more implicit type conversion fixes

+++ src/get-bbox.cpp
@@ -30,7 +30,7 @@
 
 #include "get-bbox.h"
 
-Rcpp::NumericMatrix rcpp_get_bbox (float xmin, float xmax, float ymin, float ymax)
+Rcpp::NumericMatrix rcpp_get_bbox (double xmin, double xmax, double ymin, double ymax)
 {
     std::vector <std::string> colnames, rownames;
     colnames.push_back ("min");
@@ -52,7 +52,7 @@ Rcpp::NumericMatrix rcpp_get_bbox (float xmin, float xmax, float ymin, float yma
     return bbox;
 }
 
-Rcpp::NumericVector rcpp_get_bbox_sf (float xmin, float xmax, float ymin, float ymax)
+Rcpp::NumericVector rcpp_get_bbox_sf (double xmin, double xmax, double ymin, double ymax)
 {
     std::vector <std::string> names;
     names.push_back ("xmin");
